support set firstday
