+++ test/calendar-heatmap.html
@@ -74,7 +74,7 @@
                     calendar: {
                         orient: 'vertical',
                         cellSize: 40,
-                        range: '2017-05'
+                        range: '2017-02'
                     },
                     series: {
                         type: 'effectScatter',
