+++ test/calendar-heatmapver.html
@@ -78,7 +78,7 @@
                     {
                         left: 300,
                         orient: 'vertical',
-                        range: ['2001-02-20', '2001-2-23']
+                        range: ['2001-02-20', '2001-02-23']
                     },
                     {
                         left: 530,
